basic spike trap, no art
basic spike trap. the spike trap will damage the player if they are standing on the trap for a period of time. the spike trap does not activate right away though, and the player must linger. afterward, the spike trap must take some time to rearm itself before it can damage the player again. there is no art for the spike trap yet.